Bugfix: attributes are floating points not integers

+++ src/utils.py
@@ -56,7 +56,7 @@ def create_vehicle_attrs(attrs):
     """ Returns a dictionary with the vehicle attributes """
     try:
         # Assert all values are integers, if not, discard everything
-        attribs = [int(attrib) if attrib else 0 for attrib in attrs.split(',')]
+        attribs = [float(attrib) if attrib else 0 for attrib in attrs.split(',')]
         if len(attribs) == (len(CAR_ATTRIBUTES) - 2):
             attribs.extend([1, 1])  # Adds oneway and turns restriction
             return dict(zip(CAR_ATTRIBUTES, attribs))
